+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 ### Added
 
-- Add debug logs support for world-engine token generation
+- Debug logs support for world-engine token generation
 
 ## [1.8.0] - 2023-09-20